+++ backends/simplego/gen_convgeneral_exec_full_bf16.go
@@ -51,8 +51,8 @@ func execConvBFloat16(plan convGeneralExecPlan) error { //alt:full_bf16
 	kernelDilations := params.kernelDilations                                                    //alt:full|full_bf16
 	batchGroupCount := params.batchGroupCount                                                    //alt:full|full_bf16
 	outputBatchSize := outputShape.Dimensions[inputBatchAxis]                                    //alt:full|full_bf16
-	featureGroupCount := params.featureGroupCount                                                //alt:full|full_bf16
-	numOutputChannelsPerGroup := outputShape.Dimensions[axes.OutputChannels] / featureGroupCount //alt:full|full_bf16
+	channelGroupCount := params.channelGroupCount                                                //alt:full|full_bf16
+	numOutputChannelsPerGroup := outputShape.Dimensions[axes.OutputChannels] / channelGroupCount //alt:full|full_bf16
 
 	outputBatchAxis := axes.OutputBatch
 	outputChannelsAxis := axes.OutputChannels
@@ -111,7 +111,7 @@ func execConvBFloat16(plan convGeneralExecPlan) error { //alt:full_bf16
 			// Accumulate over all the kernel/input channels.
 			inputChannelStride := inputStrides[inputChannelsAxis]
 			kernelChannelStride := kernelStrides[kernelInputChannelsAxis]
-			if featureGroupCount > 1 { //alt:full|full_bf16
+			if channelGroupCount > 1 { //alt:full|full_bf16
 				featureGroup := outputChannel / numOutputChannelsPerGroup                    //alt:full|full_bf16
 				inputFlatIdx += inputChannelStride * (featureGroup * kernelNumInputChannels) //alt:full|full_bf16
 			} //alt:full|full_bf16